integrate update project api , solve response problem  and remove language icon
integrate update project api , solve response problem  and remove language icon
