+++ src/components/layouts/ClassicLayout.tsx
@@ -22,7 +22,7 @@ const HeaderActionsStart = () => {
 const HeaderActionsEnd = () => {
     return (
         <>
-            <LanguageSelector />
+           
             <Notification />
             <SidePanel />
             <UserDropdown hoverable={false} />
