+++ src/components/layouts/DeckedLayout.tsx
@@ -22,7 +22,7 @@ const HeaderActionsEnd = () => {
     return (
         <>
             <Search />
-            <LanguageSelector />
+            
             <Notification />
             <SidePanel />
             <UserDropdown hoverable={false} />
