+++ src/components/layouts/ModernLayout.tsx
@@ -22,7 +22,6 @@ const HeaderActionsStart = () => {
 const HeaderActionsEnd = () => {
     return (
         <>
-            <LanguageSelector />
             <Notification />
             <SidePanel />
             <UserDropdown hoverable={false} />
