+++ src/components/layouts/StackedSideLayout.tsx
@@ -20,7 +20,6 @@ const HeaderActionsStart = () => {
 const HeaderActionsEnd = () => {
     return (
         <>
-            <LanguageSelector />
             <Notification />
             <SidePanel />
             <UserDropdown hoverable={false} />
