+++ src/components/template/LanguageSelector.tsx
@@ -14,9 +14,7 @@ import type { CommonProps } from '@/@types/common'
 
 const languageList = [
     { label: 'English', value: 'en', flag: 'us' },
-    { label: 'Chinese', value: 'zhCn', flag: 'cn' },
-    { label: 'Espanol', value: 'es', flag: 'sp' },
-    { label: 'Arabic', value: 'ar', flag: 'ar' },
+    
 ]
 
 const _LanguageSelector = ({ className }: CommonProps) => {
