+++ src/mock/data/authData.ts
@@ -2,8 +2,8 @@ export const signInUserData = [
     {
         id: '21',
         avatar: '/img/avatars/thumb-1.jpg',
-        userName: 'Carolyn Perkins',
-        email: 'carolyn.p@elstar.com',
+        userName: 'Devashish  ',
+        email: 'carolyn.p@.com',
         authority: ['admin', 'user'],
         password: '123Qwe',
         accountUserName: 'admin',
