+++ src/views/sales/ProductForm/ProductForm.tsx
@@ -8,7 +8,7 @@ import { Form, Formik, FormikProps, FormikProvider, useFormik } from 'formik'
 import BasicInformationFields from './BasicInformationFields'
 import PricingFields from './PricingFields'
 import OrganizationFields from './OrganizationFields'
-import ProductImages from './ProductImages'
+import ProductImages from './ProjectCreateForm'
 import cloneDeep from 'lodash/cloneDeep'
 import { HiOutlineTrash } from 'react-icons/hi'
 import { AiOutlineSave } from 'react-icons/ai'
