+++ src/views/sales/ProductNew/ProductNew.tsx
@@ -6,7 +6,7 @@ import toast from '@/components/ui/toast'
 import Notification from '@/components/ui/Notification'
 import { useNavigate } from 'react-router-dom'
 import { apiCreateSalesProduct } from '@/services/SalesService'
-import ProductImages from '@/views/sales/ProductForm/ProductImages'
+import ProductImages from '@/views/sales/ProductForm/ProjectCreateForm'
 import BasicInformationFields from '@/views/sales/ProductForm/BasicInformationFields'
 
 const ProductNew = () => {
